+++ tests/testthat/test-combine.R
@@ -1,14 +1,13 @@
 context("combine")
 
-test_that("combine handles NULL (1596)", {
+test_that("combine handles NULL (#1596, #3365)", {
   expect_equal(combine(list(NULL, 1, 2)), c(1, 2))
   expect_equal(combine(list(1, NULL, 2)), c(1, 2))
   expect_equal(combine(list(1, 2, NULL)), c(1, 2))
-  expect_error(
-    combine(list(NULL, NULL)),
-    "no data to combine, all elements are NULL",
-    fixed = TRUE
-  )
+  expect_equal(combine(), logical())
+  expect_equal(combine(list(NULL)), logical())
+  expect_equal(combine(list(NULL, NULL), list(NULL)), list(NULL, NULL, NULL))
+  expect_equal(combine(NULL, list(NULL, NULL)), list(NULL, NULL))
 })
 
 test_that("combine complains about incompatibilites", {